Support different volumes in tls-cert-manager
The current script of the tls cert manager does not allow mounting the zimbra instance into an own volume.

With the following mounts in Kubernetes, the tls-update-manager fails:
```
        - name: zimbra-os
          mountPath: /data
        - name: zimbra-root
          mountPath: /data/opt/zimbra
```

This is due to os.replace not supporting moving data from one mounted volume to another.

Replacing it with `shutil.copy2` does support this edge case.

Signed-off-by: Christoph Brand <christoph@xcnt.io>